add logging interceptor, DateUtils minor changes

+++ app/src/main/java/com/bks/eurosporttest/data/network/model/StoryDto.kt
@@ -18,7 +18,7 @@ data class StoryDto(
     var image: String,
 
     @SerializedName("date")
-    var date: Long,
+    var date: Double,
 
     @SerializedName("author")
     var author: String,
